requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
async-timeout==4.0.2
attrs==22.2.0
bracex==2.4
cloudpickle==1.6.0
contourpy==1.0.7
cycler==0.11.0
einops==0.6.0
faiss-gpu==1.7.2
numpy==1.23.5
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.3.0
h5py==2.10.0
huggingface-hub==0.12.1
imageio==2.26.0
importlib-resources==5.12.0
joblib==1.2.0
kapture==1.1.8
kiwisolver==1.4.4
kornia==0.6.9
lightning-utilities==0.8.0
llvmlite==0.41.1
loguru==0.6.0
matplotlib==3.7.1
multidict==6.0.4
networkx==3.0
numba==0.58.1
numpy-quaternion==2022.4.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
piexif==1.1.3
Pillow==9.4.0
plyfile==1.0.1
protobuf==3.20.3
pycolmap==0.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.9.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
scikit-image==0.19.1
scikit-learn==1.2.2
scipy==1.10.1
tabulate==0.9.0
tensorboardX==2.6
threadpoolctl==3.1.0
tifffile==2023.2.28
timm==0.6.11
torchmetrics==0.11.4
tqdm==4.65.0
wcmatch==8.5
yacs==0.1.8
yarl==1.8.2
zipp==3.15.0
